refactor: course main to widgets
