+++ src/pages/angular.tsx
@@ -4,12 +4,12 @@ import { AngularForMentors } from '@/features/angular-for-mentors';
 import { AngularTopics } from '@/features/angular-topics';
 import { Certification } from '@/features/certification';
 import { Communication } from '@/features/communication';
-import { CourseMain } from '@/features/course-main';
 import { Required } from '@/features/required';
 import { StudyPath } from '@/features/study-path';
 import { Trainers } from '@/features/trainers';
 import { angular } from '@/features/trainers/angular.data';
 import { TrainingProgram } from '@/features/training-program';
+import { CourseMain } from '@/widgets/course-main';
 
 const COURSE_NAME = 'angular';
