+++ src/pages/aws-fundamentals.tsx
@@ -2,11 +2,11 @@ import { Breadcrumbs } from '@/app/components';
 import { About } from '@/features/about';
 import { Certification } from '@/features/certification';
 import { Communication } from '@/features/communication';
-import { CourseMain } from '@/features/course-main';
 import { Required } from '@/features/required';
 import { Trainers } from '@/features/trainers';
 import { awsFundamentals } from '@/features/trainers/awsFundamentals.data.ts';
 import { TrainingProgram } from '@/features/training-program';
+import { CourseMain } from '@/widgets/course-main';
 
 const COURSE_NAME = 'aws fundamentals';
