+++ src/pages/javascript-preschool-ru.tsx
@@ -2,12 +2,12 @@ import { Breadcrumbs } from '@/app/components';
 import { About } from '@/features/about';
 import { Certification } from '@/features/certification';
 import { Communication } from '@/features/communication';
-import { CourseMain } from '@/features/course-main';
 import { Faq } from '@/features/faq';
 import { Required } from '@/features/required';
 import { Trainers } from '@/features/trainers';
 import { preSchoolRu } from '@/features/trainers/preSchool.data.ts';
 import { TrainingProgram } from '@/features/training-program';
+import { CourseMain } from '@/widgets/course-main';
 
 const COURSE_NAME = 'js / front-end pre-school ru';
