+++ src/pages/nodejs.tsx
@@ -3,11 +3,11 @@ import { Breadcrumbs } from '@/app/components';
 import { About } from '@/features/about';
 import { Certification } from '@/features/certification';
 import { Communication } from '@/features/communication';
-import { CourseMain } from '@/features/course-main';
 import { Required } from '@/features/required';
 import { Trainers } from '@/features/trainers';
 import { nodejs } from '@/features/trainers/nodejs.data.ts';
 import { TrainingProgram } from '@/features/training-program';
+import { CourseMain } from '@/widgets/course-main';
 
 const COURSE_NAME = 'node.js';
 export const Nodejs: FC = () => {
