+++ src/pages/react.tsx
@@ -2,10 +2,10 @@ import { Breadcrumbs } from '@/app/components';
 import { About } from '@/features/about';
 import { Certification } from '@/features/certification';
 import { Communication } from '@/features/communication';
-import { CourseMain } from '@/features/course-main';
 import { Trainers } from '@/features/trainers';
 import { reactEn } from '@/features/trainers/react-en.data.ts';
 import { TrainingProgram } from '@/features/training-program';
+import { CourseMain } from '@/widgets/course-main';
 
 const COURSE_NAME = 'react';
 export const React = () => {
