+++ src/widgets/course-main/course-main.module.scss
