+++ src/widgets/course-main/course-main.test.tsx
