+++ src/widgets/course-main/course-main.tsx
