+++ src/widgets/course-main/courseLoader.ts
