+++ src/widgets/course-main/index.ts
