+++ src/widgets/course-main/utils/get-course-status.ts
