+++ src/widgets/course-main/utils/has-day.ts
