+++ src/widgets/course-main/utils/index.ts
